requirements.txt
# ── ORIA Flask Backend — Pinned Dependencies (S-05) ─────────────────────────
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-WTF>=1.2,<2.0
openai>=1.50,<2.0
python-dotenv>=1.0,<2.0
Werkzeug>=3.0,<4.0
gunicorn>=22.0,<23.0

# ── ORIA Telegram Bot ───────────────────────────────────────────────────────
aiogram>=3.4,<4.0
aiohttp>=3.9,<4.0
